tests/wpt/tests/graphics-aria/graphics-roles.html

<!doctype html>
<html>
<head>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/resources/testdriver.js"></script>
  <script src="/resources/testdriver-vendor.js"></script>
  <script src="/resources/testdriver-actions.js"></script>
  <script src="/wai-aria/scripts/aria-utils.js"></script>
</head>
<body>

<p>Tests the concrete roles defined in the <a href="https://www.w3.org/TR/graphics-aria-1.0/#role_definitions">ARIA Graphics Module</a> role definitions.</p>

<div role="graphics-document" data-expectedrole="graphics-document" class="ex">x</div>
<div role="graphics-object" data-expectedrole="graphics-object" class="ex">x</div>
<div role="graphics-symbol" data-expectedrole="graphics-symbol" class="ex">x</div>

<script>
AriaUtils.verifyRolesBySelector(".ex");
</script>

</body>
</html>
