LayoutTests/imported/w3c/web-platform-tests/graphics-aria/graphics-roles-expected.txt

Tests the concrete roles defined in the ARIA Graphics Module role definitions.

x
x
x

FAIL graphics-document assert_equals: <div role="graphics-document" data-expectedrole="graphics-document" class="ex">x</div> expected "graphics-document" but got "document"
FAIL graphics-object assert_equals: <div role="graphics-object" data-expectedrole="graphics-object" class="ex">x</div> expected "graphics-object" but got "group"
FAIL graphics-symbol assert_equals: <div role="graphics-symbol" data-expectedrole="graphics-symbol" class="ex">x</div> expected "graphics-symbol" but got "image"
